Add data loading goodies
We need to load the MNIST dataset in a simple-yet-uncommon
(at least that we're aware of) format. The format
itself is documented on `http://yann.lecun.com/exdb/mnist/`
and is basically a leading bit of metadata followed by raw
data points. Given MNIST is a collection of handwritten digits,
we've also added some functions generating PNG images for each
of the entries so as to visually inspect the raw data itself.