Add XOR data generation functions
These files provide an easy way of generating XOR-based
data. Its distribution in a 2D space is very handy to
check whether the MLP is behaving as expected. The 4
well-defined clusters that end up forming allow a human
(i.e. us) to have a very intuitive feedback as to what's
going on behind the scenes.

In order to avoid the entire dataset to be purely synthetic,
we're also adding some normal noise to the basic data points.