Add an initial (and unfinished) MLP implementation
We've implemented a MLP that's capable of initialising and
computing activations on each layer for now. We still have
to implement the backpropagation algorithm allowing it to
actually 'learn'.

A thing to note is the implementation of the `String()` method,
which allows us to print out relevant MLP information with a
call such as: `fmt.Printf("\n%s", mlp_instance)`.