Add MLP tests and backing data
We've designed some tests to check whether forward propagation
(i.e. the computing of activations at each layer) was behaving
as expected. In order to do so, we've had to serialise and
de-serialise the results so as to be able to make correct
comparisons between the matrices describing the MLPs state.