Remove the need for passing matrix indices
Gonum requires a `func(int, int, float64) float64` when
calling the `Apply()` method on a matrix. However, this
should be hidden from the end user: an activation function
shouldn't be concerned with the indices within the matrix
of the value it's acting on. We've added a wrapper when
calling `Apply()` to circumvent this issue.